move the steps and schema out of the lib subfolder
