+++ libs/newsletter-workflow/src/schemas.ts
