+++ libs/newsletter-workflow/src/steps/brazeLayout.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import type { WizardStepLayout } from '@newsletters-nx/state-machine';
-import { executeModify } from '../executeModify';
+import { executeModify } from '../lib/executeModify';
 
 const markdownToDisplay = `
 # Modify Braze Values
