+++ libs/newsletter-workflow/src/steps/descriptionLayout.ts
@@ -3,7 +3,7 @@ import type {
 	WizardStepData,
 	WizardStepLayout,
 } from '@newsletters-nx/state-machine';
-import { executeModify } from '../executeModify';
+import { executeModify } from '../lib/executeModify';
 
 const markdownToDisplay = `
 # Specify the Description
