+++ libs/newsletter-workflow/src/steps/finishLayout.ts
@@ -1,6 +1,6 @@
 import type { WizardStepLayout } from '@newsletters-nx/state-machine';
-import { getStringValuesFromRecord } from '../getValuesFromRecord';
-import { regExPatterns } from '../regExPatterns';
+import { getStringValuesFromRecord } from '../lib/getValuesFromRecord';
+import { regExPatterns } from '../lib/regExPatterns';
 
 const markdownTemplate = `# Finished
