+++ libs/newsletter-workflow/src/steps/pillarLayout.ts
@@ -4,9 +4,9 @@ import type {
 	WizardStepData,
 	WizardStepLayout,
 } from '@newsletters-nx/state-machine';
-import { executeModify } from '../executeModify';
-import { getStringValuesFromRecord } from '../getValuesFromRecord';
-import { regExPatterns } from '../regExPatterns';
+import { executeModify } from '../lib/executeModify';
+import { getStringValuesFromRecord } from '../lib/getValuesFromRecord';
+import { regExPatterns } from '../lib/regExPatterns';
 
 const markdownTemplate = `
 # Select a Pillar for {{name}}
